+++ NEWS.rst
@@ -26,6 +26,8 @@ NEWS for system-image updater
    better with configuration reloading.  (LP: #1412698)
  * Support for the ``/etc/ubuntu-build`` file has been removed.  The build
    number now comes from the configuration files.  (LP: #1377312)
+ * Move the ``archive-master.tar.xz`` file to ``/usr/share/system-image`` for
+   better FHS compliance.  (LP: #1377184)
 
 2.5.1 (2014-10-21)
 ==================